fix: revise wording
