--- src/app/card-id-consent/page.tsx
@@ -1,5 +1,0 @@
-import VerifyIdContainer from "@/features/card-id-consent/containers/VerifyIdContainer";
-
-export default function VerifyPage() {
-  return <VerifyIdContainer />;
-}
