+++ src/app/id-consent/page.tsx
@@ -1,0 +1,5 @@
+import VerifyIdContainer from "@/features/id-consent/containers/VerifyIdContainer";
+
+export default function VerifyPage() {
+  return <VerifyIdContainer />;
+}
